requirements.txt
arrow==1.2.3
google_api_python_client==2.86.0
icalendar==5.0.5
protobuf==4.23.0
psycopg2_binary==2.9.6
python_dateutil==2.8.2
python_gitlab==3.14.0
pytz==2023.3
PyYAML==6.0
tqdm==4.65.0
